requirements.txt
PyNaCl
pyparsing
yt-dlp
discord.py
ffmpeg
aiohttp
requests
async-timeout
google-api-python-client
psutil
humanize
